move exporter to conftest
