+++ tests/unit/test_open_telemetry.py
@@ -8,10 +8,6 @@ import os
 from unittest import mock
 
 import pytest
-from opentelemetry import trace
-from opentelemetry.sdk.resources import SERVICE_NAME, Resource
-from opentelemetry.sdk.trace import TracerProvider
-from opentelemetry.sdk.trace.export import SimpleSpanProcessor
 from opentelemetry.sdk.trace.export.in_memory_span_exporter import InMemorySpanExporter
 from opentelemetry.trace import span
 
@@ -82,17 +78,6 @@ api_calls = [
     {"name": "Session.create_dataframe[values]"},
     {"name": "DataFrame.to_df", "subcalls": [{"name": "DataFrame.select"}]},
 ]
-
-
-@pytest.fixture(scope="module")
-def dict_exporter():
-    resource = Resource(attributes={SERVICE_NAME: "snowpark-python-open-telemetry"})
-    trace_provider = TracerProvider(resource=resource)
-    dict_exporter = InMemorySpanExporter()
-    processor = SimpleSpanProcessor(dict_exporter)
-    trace_provider.add_span_processor(processor)
-    trace.set_tracer_provider(trace_provider)
-    yield dict_exporter
 
 
 def test_without_open_telemetry(monkeypatch, dict_exporter):
